move `composerKeywordsToYouTube` closer to where its being used
